requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
altair
google-api-python-client
google-auth
